+++ src/MarginTrading.Backend.Contracts/Positions/PositionDirectionContract.cs
@@ -1,8 +1,21 @@
-﻿namespace MarginTrading.Backend.Contracts.Positions
+﻿using JetBrains.Annotations;
+
+namespace MarginTrading.Backend.Contracts.Positions
 {
+    /// <summary>
+    /// The direction of a position
+    /// </summary>
+    [PublicAPI]
     public enum PositionDirectionContract
     {
+        /// <summary>
+        /// Position is profitable if the price goes up 
+        /// </summary>
         Long = 1,
+        
+        /// <summary>
+        /// Position is profitable if the price goes down
+        /// </summary>
         Short = 2
     }
 }
