+++ src/MarginTrading.Backend.Contracts/Prices/BestPriceContract.cs
