--- src/MarginTrading.Backend.Contracts/Snow/Orders/OrderDirectionContract.cs
@@ -1,18 +1,0 @@
-﻿namespace MarginTrading.Backend.Contracts.Snow.Orders
-{
-    /// <summary>
-    /// The direction of an order
-    /// </summary>
-    public enum OrderDirectionContract
-    {
-        /// <summary>
-        /// Order to buy the quoting asset of a pair
-        /// </summary>
-        Buy = 1,
-        
-        /// <summary>
-        /// Order to sell the quoting asset of a pair
-        /// </summary>
-        Sell = 2
-    }
-}
