--- src/MarginTrading.Backend.Contracts/Snow/Orders/OriginatorTypeContract.cs
@@ -1,9 +1,0 @@
-﻿namespace MarginTrading.Backend.Contracts.Snow.Orders
-{
-    public enum OriginatorTypeContract
-    {
-        Investor = 1,
-        System = 2,
-        OnBehalf = 3
-    }
-}
