+++ src/MarginTrading.Backend/Controllers/PricesController.cs
@@ -1,13 +1,15 @@
 ﻿using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using MarginTrading.Backend.Attributes;
+using MarginTrading.Backend.Contracts;
+using MarginTrading.Backend.Contracts.Prices;
 using MarginTrading.Backend.Contracts.Snow.Prices;
 using MarginTrading.Backend.Core;
-using MarginTrading.DataReader.Services;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 
-namespace MarginTrading.DataReader.Controllers.Snow
+namespace MarginTrading.Backend.Controllers
 {
     /// <summary>                                                                                       
     /// Provides data about prices
@@ -16,11 +18,11 @@ namespace MarginTrading.DataReader.Controllers.Snow
     [Route("api/prices")]
     public class PricesController : Controller, IPricesApi
     {
-        private readonly IQuotesSnapshotReadersService _quotesSnapshotReadersService;
+        private readonly IQuoteCacheService _quoteCacheService;
 
-        public PricesController(IQuotesSnapshotReadersService quotesSnapshotReadersService)
+        public PricesController(IQuoteCacheService quoteCacheService)
         {
-            _quotesSnapshotReadersService = quotesSnapshotReadersService;
+            _quoteCacheService = quoteCacheService;
         }
 
         /// <summary>
@@ -29,14 +31,20 @@ namespace MarginTrading.DataReader.Controllers.Snow
         /// <remarks>
         /// Post because the query string will be too long otherwise
         /// </remarks>
-        [HttpGet, Route("best")]
-        public async Task<List<BestPriceContract>> Best(string[] assetPairsIds)
+        [Route("best")]
+        [HttpPost]
+        [SkipMarginTradingEnabledCheck]
+        public async Task<Dictionary<string, BestPriceContract>> GetBestAsync(
+            [FromBody] InitPricesBackendRequest request)
         {
-            var dict = _quotesSnapshotReadersService.GetSnapshotAsync();
-            return assetPairsIds.Select(pid => dict.GetValueOrDefault(pid)).Where(p => p != null).Select(Convert)
-                .ToList();
-        }
+            IEnumerable<KeyValuePair<string, InstrumentBidAskPair>> allQuotes = _quoteCacheService.GetAllQuotes();
 
+            if (request.AssetIds != null && request.AssetIds.Any())
+                allQuotes = allQuotes.Where(q => request.AssetIds.Contains(q.Key));
+
+            return allQuotes.ToDictionary(q => q.Key, q => Convert(q.Value));
+        }
+        
         private BestPriceContract Convert(InstrumentBidAskPair arg)
         {
             return new BestPriceContract
